requirements.txt
PyJWT
uuid
requests
pandas